correctly free data allocated by PVerFileParser
Adjust return type of ver_iter_* to be *const c_char, as that data does
not need manual freeing.

ver_file_parser_* function do require it (they use to_c_string) and the
parser itself also needs to be freed. Add a SafeVerFileParser type to
free the parser on drop, and free the data returned by the parser ASAP
after cloning into a Rust String.

Data is allocated with C++ 'new', so we add custom functions to call
'delete' instead of relying on libc::free to avoid any potential issues.